Minor fixes for Linux version
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@930 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Logic/BluRaySup/BluRaySupPicture.cs
@@ -619,7 +619,7 @@ namespace Nikse.SubtitleEdit.Logic.BluRaySup
                 ToolBox.SetWord(headerPCSStart, 2, h);                              // cropped height
                 ToolBox.SetByte(headerPCSStart, 4, fpsId);
                 ToolBox.SetWord(headerPCSStart, 5, pic.CompositionNumber);
-                headerPCSStart[14] = (byte)(pic.IsForced ? (byte)0x40 : 0);
+                headerPCSStart[14] = (byte)(pic.IsForced ? 0x40 : 0);
                 ToolBox.SetWord(headerPCSStart, 15, pic.WindowXOffset);
                 ToolBox.SetWord(headerPCSStart, 17, yOfs);
                 for (int i = 0; i < headerPCSStart.Length; i++)
